+++ tests/TextUI/dataprovider-issue-2859.phpt
